fix ramp_in
There are a couple of problems with the relatively new ramp_in feature
of fio.  First, the estimated time to completion did not correctly take
it into account and bounces around.  Second and more importantly, the
runtime was including ramp in time in throughput calculations even
though the IO done during that time was ignored, thus making throughput
metrics incorrect. This patch fixes both.

+++ fio.h
@@ -126,7 +126,7 @@ struct thread_stat {
 	unsigned long total_complete;
 
 	unsigned long long io_bytes[2];
-	unsigned long runtime[2];
+	unsigned long long runtime[2];
 	unsigned long total_run_time;
 
 	/*
